backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.12.5
pydantic-settings==2.4.0

